[m3] Move Collatz code into a test file
